+++ solutions/devsprint-bruno-ramos-3/DeliveryApp/Protocols/ViewCode.swift
@@ -5,4 +5,21 @@
 //  Created by Rodrigo Francischett Occhiuto on 14/02/22.
 //
 
-import Foundation
+import UIKit
+
+protocol ViewCode {
+    func setupComponents()
+    func setupConstraints()
+    func setupExtraConfiguration()
+    func setup()
+}
+
+extension ViewCode {
+    func setup() {
+        setupComponents()
+        setupConstraints()
+        setupExtraConfiguration()
+    }
+
+    func setupExtraConfiguration() {}
+}
